Split for code sharing
